feat(mixpanel): Add `context` mixpanel plugin config option
- Move the mixpanel load script to separate file
- Modify the load script to be able to specify which context object
the mixpanel instance should be loaded. Defaults to `window` for
backward compatibility
- Add `context` mixpanel plugin config option that specifies the context
that's passed the refactored load script
- Update hooks to get mixpanel instance from resolveMixpanel function
instead of from `window`. This decouples the use of mixpanel instance
from the window object
